+++ service_location/src/main/kotlin/org/thechance/service_location/plugins/StatusExceptions.kt
@@ -2,7 +2,7 @@ package org.thechance.service_location.plugins
 
 import io.ktor.server.application.*
 import io.ktor.server.plugins.statuspages.*
-import org.thechance.service_location.data.utils.configureStatusPages
+import org.thechance.service_location.util.configureStatusPages
 
 fun Application.configureStatusExceptions() {
     install(StatusPages) {
